Clean up home page by extraction styles to a mixin
